+++ src/models/StyledComponent.js
@@ -95,7 +95,7 @@ export default (ComponentStyle: Function) => {
         propsForElement.className = [className, generatedClassName].filter(x => x).join(' ')
         if (innerRef) {
           propsForElement.ref = innerRef
-          delete propsForElement.innerRef
+          if (isTag(target)) delete propsForElement.innerRef
         }
 
         return createElement(target, propsForElement, children)